Simplify type of BibTeX parser function
So far, BibTeX.Parser.parse was designed to return the remaining input stream
after parsing has finished, together with the actual Database. When you looked
carefully at the function, you could notice, that the remaining stream returned
by the function could not be anything else but the empty String (i.e. ""). So,
this return value is not really needed.