Total remake and extension of the tag value parsing system
BibTeX.Parser.parseTagValue has been replaced by a call to the new higher-order
BibTeX.Parser._parseTagValue, which takes a list of different value parsers as
an argument. For each parser, _parseTagValue checks, if this parser can be
applied to the input stream. The first matching parser is then used to read the
tag value from the stream. This plug-in system makes it easier to see the
different parsing mechanisms, and helps when changing them in the future. It
also removes a lot of redundancy in the code.

The value parser now supports string references and string concatenation.