Make string declarations flat in BibTeX-XML
Because the previously used stringDeclaration elements don't hold any data
other than the single string definitions themselves, they are not useful and
only blow up the database. The decision to remove these elements and make
string declarations flat was made during writing the first part of the project
documentation.

To still make use of Haskell's type classes, there were a bunch of lines in
which a simple type had to be wrapped inside a list. This resulted in lots of
brackets which needed to be added all over the code of
bibtoxml/BibTeX/Output/XML.hs and therefore a lot of indentation changes.
This pushed the count of changed lines to the sky while the actual amount of
change wasn't that big.