Make the parser strict and implement error recovery
Previously, the parser wasn't strict in turns of entry types and field names.
If a type or name that was not a defined BibTeX identifier had occured in a
database, the parser would still have read it without raising an error and
kept it as a String until output. Such wrong identifiers would have ended up
in, for example, a BibTeX-XML document which has a very strict definition of
allowed values for entry types or field names. Because of this, the output
generated by the BibTeX library had the potential to create wrong output.

This commit introduces a Result type inside the parser to return either a
Success with a result value, a remaining input stream and a log buffer (a List
of Strings) or a Failure which only has a log buffer. In case of an error while
parsing a database, a parser function can now tell the calling function that
something went wrong, without breaking the whole program down immediately.

The parser now drops all entries (or at least it gives its best to try to)
which have a wrong entry type or field name.

The log list is a convenient feature to capture information about the whole
parsing process~~~not just the errors. But currently, it is only used for
errors. Also, this commit does not make the whole parser robust but only a few
functions of it. This is to save time because we are in a hurry to finish the
submission for the exam.